# Conflicts:
#	SP2/Common/Debug/Common.log
#	SP2/Common/Debug/Common.tlog/CL.read.1.tlog
#	SP2/Common/Debug/Common.tlog/CL.write.1.tlog
#	SP2/Common/Debug/Common.tlog/Common.lastbuildstate
#	SP2/Common/Debug/Common.tlog/Lib-link.read.1.tlog
#	SP2/Common/Debug/Common.tlog/Lib-link.write.1.tlog
#	SP2/Common/Debug/Common.tlog/cl.command.1.tlog
#	SP2/Common/Debug/Common.tlog/lib.command.1.tlog
#	SP2/Common/Debug/MatrixStack.obj
#	SP2/Common/Debug/Mtx44.obj
#	SP2/Common/Debug/Vector3.obj
#	SP2/Common/Debug/timer.obj
#	SP2/Common/Debug/vc120.idb
#	SP2/Common/Debug/vc120.pdb
#	SP2/Debug/Common.lib
#	SP2/Debug/LiZhaoyuan1.exe
#	SP2/Debug/LiZhaoyuan1.ilk
#	SP2/Debug/LiZhaoyuan1.pdb
#	SP2/LiZhaoyuan1/Debug/Application.obj
#	SP2/LiZhaoyuan1/Debug/Camera.obj
#	SP2/LiZhaoyuan1/Debug/Camera2.obj
#	SP2/LiZhaoyuan1/Debug/Camera3.obj
#	SP2/LiZhaoyuan1/Debug/Camera4.obj
#	SP2/LiZhaoyuan1/Debug/LiZhaoyuan1.log
#	SP2/LiZhaoyuan1/Debug/LiZhaoyuan1.tlog/CL.read.1.tlog
#	SP2/LiZhaoyuan1/Debug/LiZhaoyuan1.tlog/CL.write.1.tlog
#	SP2/LiZhaoyuan1/Debug/LiZhaoyuan1.tlog/LiZhaoyuan1.lastbuildstate
#	SP2/LiZhaoyuan1/Debug/LiZhaoyuan1.tlog/cl.command.1.tlog
#	SP2/LiZhaoyuan1/Debug/LiZhaoyuan1.tlog/link.command.1.tlog
#	SP2/LiZhaoyuan1/Debug/LiZhaoyuan1.tlog/link.read.1.tlog
#	SP2/LiZhaoyuan1/Debug/LiZhaoyuan1.tlog/link.write.1.tlog
#	SP2/LiZhaoyuan1/Debug/LoadOBJ.obj
#	SP2/LiZhaoyuan1/Debug/LoadTGA.obj
#	SP2/LiZhaoyuan1/Debug/Mesh.obj
#	SP2/LiZhaoyuan1/Debug/MeshBuilder.obj
#	SP2/LiZhaoyuan1/Debug/RenderGalaxy.obj
#	SP2/LiZhaoyuan1/Debug/RenderMun.obj
#	SP2/LiZhaoyuan1/Debug/RenderSoraJewel.obj
#	SP2/LiZhaoyuan1/Debug/SP2.obj
#	SP2/LiZhaoyuan1/Debug/SceneGalaxy.obj
#	SP2/LiZhaoyuan1/Debug/SceneMun.obj
#	SP2/LiZhaoyuan1/Debug/SceneSoraJewel.obj
#	SP2/LiZhaoyuan1/Debug/Utility.obj
#	SP2/LiZhaoyuan1/Debug/main.obj
#	SP2/LiZhaoyuan1/Debug/shader.obj
#	SP2/LiZhaoyuan1/Debug/vc120.idb
#	SP2/LiZhaoyuan1/Debug/vc120.pdb
#	SP2/LiZhaoyuan1/Source/Camera3.cpp

+++ SP2/LiZhaoyuan1/Source/SceneGalaxy.h
@@ -160,9 +160,9 @@ private:
 
 	float boxX = 0;
 	float boxZ = 0;
-	int camPosX = 0;
-	int camPosY = 0;
-	int camPosz = 0;
+	float camPosX = 0;
+	float camPosY = 0;
+	float camPosz = 0;
 
 	void MunInit();
 	void MunRender();
